Move Timestamp and related classes into apache_beam/utils.
